pyproject.toml
[project]
name = "resume-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "nltk>=3.9.1",
    "pdfminer-six>=20250327",
    "python-docx>=1.1.2",
    "sentence-transformers>=4.0.2",
    "torch>=2.6.0",
]
